backend/requirements.txt
fastapi
uvicorn
nltk
scikit-learn
pdfplumber
python-docx
textract
gensim
